SCALE-SEAM (86): a shared word is not a shared domain, for the third slice running
Six methods out of client.ts (1,173 -> 1,131) into three existing mixins.

  is every job on track?          executivePortfolio,
                                  constructionPortfolio -> evm.ts
  which projects do we            portfolioPrioritization -> proforma.ts
  prioritise?
  save a selection rule and       smartViews, smartViewsSave,
  run it                          smartViewRun -> elements.ts

THE PORTFOLIO TRIO SPLIT 2/1, AND ONLY THE SHAPES SAID SO. All three are
/portfolio/* global rollups, so any route-based cut moves them together.
But evm.ts recorded at ⓫ that projectHealth answers "is the job on track
across domains?" for ONE project - and executivePortfolio is that same
question over the whole portfolio, constructionPortfolio the same one
domain narrower. Those two REPORT. portfolioPrioritization ranks projects
by a weighted composite of return, budget, schedule and risk and names a
best and worst: it DECIDES, which is a pipeline question, not a status one.

⓫ also recorded that "safety sat below and did not come", and
constructionPortfolio does carry recordables. It came anyway - a
cross-project execution rollup that happens to include a safety count, not
the safety module. The tension is named in the header rather than glossed:
a slice that quietly contradicts an earlier one's stated reason is how the
next reader stops trusting either.

THE THIRD HOMONYM IN THREE SLICES. smartViews nearly went to modules.ts,
and the name matched EXACTLY - that mixin already owns a saved-views
family, SavedViewDef at /projects/{pid}/modules/{key}/views. Comparing the
two types separated them in one line. A SavedViewDef is {q, state, sort}, a
data-grid filter over a register. A SmartView is {selector, mode:
isolate|color|hide, color} and smartViewRun returns matched GUIDs. Same
word, unrelated things.

Three in three slices: rvtBridgeStatus versus the LAND-USE entitlements.ts
at (85), "they are all multipart uploads" as a question at (85), and now
"view". A name is evidence about vocabulary, not about domain - and in a
codebase spanning development finance, construction operations and 3D
geometry, the same word reliably means three things.

SmartView moved out of client.ts's type imports into elements.ts's.

Verification, now including the three gates (85) missed: tsc --noEmit exit
0; full web suite 203 files / 2,059 tests; eslint clean; unfiledMap gate
green at 17; test_file_sizes pin 1,173 -> 1,131 and exact;
test_roadmap_status, test_claude_md_gates, test_changelog_current,
test_reachable, test_declared_imports all pass; ruff clean.
